pyproject.toml
# pytest settings that might help... feel free to modify or remove as needed

[project]
name = "fake-google"
version = "0.1.0"
description = "A fake Google search engine demonstrating AI hallucinations"
requires-python = ">=3.8"
dependencies = [
    "flask",
    "python-dotenv",
    "openai",
    "argparse",
    "pymongo>=4.6.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
minversion = "7.0" # minimum pytest version
addopts = "-ra -q" # default pytest command line options
pythonpath = [
  "."
]
testpaths = [
    "tests",
]
